+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import TodoCount from './TodoCount';
+import Filter from '../Filter/';
+import TodoClearButton from './TodoClearButton';
+
+const Footer = () => {
+  return (
+    <footer className="footer">
+      <TodoCount />
+      <Filter />
+      <TodoClearButton />
+    </footer>
+  );
+};
+
+export default Footer;
