+++ src/components/Footer/TodoCount.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 const TodoCount = () => {
   return (
     <span className="todo-count">
